EDC/requirements.txt
pandas==0.25.3
scikit-learn==0.22
paho-mqtt
cryptography==2.6.1
asymcrypt
